+++ src/src.rs
@@ -7,12 +7,14 @@
 
 #![deny(missing_docs)]
 
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::fmt;
+use std::iter::FromIterator;
 use std::path::Path;
 
 use serde::ser::{Serialize, Serializer};
 
+use crate::sess::Session;
 use crate::target::{TargetSet, TargetSpec};
 
 /// A source file group.
@@ -26,6 +28,8 @@ pub struct SourceGroup<'ctx> {
     pub target: TargetSpec,
     /// The directories to search for include files.
     pub include_dirs: Vec<&'ctx Path>,
+    /// The directories exported by dependent package for include files.
+    pub export_incdirs: HashMap<String, Vec<&'ctx Path>>,
     /// The preprocessor definitions.
     pub defines: HashMap<&'ctx str, Option<&'ctx str>>,
     /// The files in this group.
@@ -93,12 +97,121 @@ impl<'ctx> SourceGroup<'ctx> {
                 independent: self.independent,
                 target: self.target.clone(),
                 include_dirs: self.include_dirs.clone(),
+                export_incdirs: self.export_incdirs.clone(),
                 defines: self.defines.clone(),
                 files: files,
                 dependencies: self.dependencies.clone(),
             }
             .simplify(),
         )
+    }
+
+    /// Recursively get dependency names.
+    fn get_deps(&self, packages: &HashSet<String>, excludes: &HashSet<String>) -> HashSet<String> {
+        let mut result = packages.clone();
+
+        match self.package {
+            Some(x) => {
+                if result.contains(x) {
+                    result.extend(HashSet::<String>::from_iter(self.dependencies.clone()));
+                    result = &result - &excludes;
+                }
+            }
+            None => {}
+        }
+
+        for file in &self.files {
+            match file {
+                SourceFile::Group(group) => {
+                    result.extend(group.get_deps(&result, excludes));
+                }
+                _ => {}
+            }
+        }
+
+        result
+    }
+
+    /// Get list of packages based on constraints.
+    pub fn get_package_list(
+        &self,
+        sess: &Session,
+        packages: &HashSet<String>,
+        excludes: &HashSet<String>,
+        no_deps: bool,
+    ) -> HashSet<String> {
+        let mut result = HashSet::new();
+
+        if !packages.is_empty() {
+            result.extend(packages.clone());
+        } else {
+            result.insert(sess.manifest.package.name.to_string());
+        }
+
+        result = &result - &excludes;
+
+        if !no_deps {
+            let mut curr_length = 0;
+            while curr_length < result.len() {
+                curr_length = result.len();
+                result.extend(self.get_deps(&result, excludes));
+            }
+        }
+
+        result
+    }
+
+    /// Filter the sources, keeping only the ones that apply to the selected packages.
+    pub fn filter_packages(&self, packages: &HashSet<String>) -> Option<SourceGroup<'ctx>> {
+        let mut files = Vec::new();
+
+        if self.package.is_none() || packages.contains(self.package.unwrap()) {
+            files = self
+                .files
+                .iter()
+                .filter_map(|file| match *file {
+                    SourceFile::Group(ref group) => group
+                        .filter_packages(packages)
+                        .map(|g| SourceFile::Group(Box::new(g))),
+                    ref other => Some(other.clone()),
+                })
+                .collect();
+        }
+
+        let mut export_incdirs = self.export_incdirs.clone();
+        export_incdirs.retain(|k, _| packages.contains(k));
+        Some(
+            SourceGroup {
+                package: self.package,
+                independent: self.independent,
+                target: self.target.clone(),
+                include_dirs: self.include_dirs.clone(),
+                export_incdirs: export_incdirs,
+                defines: self.defines.clone(),
+                files: files,
+                dependencies: self.dependencies.clone(),
+            }
+            .simplify(),
+        )
+    }
+
+    /// Return list of unique include directories for the current src
+    pub fn get_incdirs(self) -> Vec<&'ctx Path> {
+        let dep_incdirs: Vec<&Path> = self
+            .export_incdirs
+            .into_iter()
+            .map(|(_, v)| v)
+            .flat_map(|it| it.clone())
+            .collect();
+        let incdirs = self
+            .include_dirs
+            .into_iter()
+            .chain(dep_incdirs.into_iter())
+            .fold(HashSet::new(), |mut acc, inc_dir| {
+                acc.insert(inc_dir);
+                acc
+            });
+        incdirs.into_iter().collect()
     }
 
     /// Flatten nested source groups.
@@ -142,12 +255,14 @@ impl<'ctx> SourceGroup<'ctx> {
                             .map(|&i| i.clone())
                             .collect(),
                     );
-                    grp.include_dirs = self
-                        .include_dirs
-                        .iter()
-                        .cloned()
-                        .chain(grp.include_dirs.into_iter())
-                        .collect();
+                    grp.include_dirs = HashSet::<&Path>::from_iter(
+                        self.include_dirs
+                            .iter()
+                            .cloned()
+                            .chain(grp.include_dirs.into_iter()),
+                    )
+                    .into_iter()
+                    .collect();
                     grp.defines = self
                         .defines
                         .iter()